requirements.txt
requests
pandas
pytest
apache-airflow
pyarrow